chore(client): move projects components
